Data definition (#43)
* pull req suggestions
* check_number rewrite
* data_def_type constructors refactor
* refactoring
* added reference
* reverting package-lock change
* nullopt fix
* set location counter for DC
* location counter in address fix
* UB fix
* refactoring and fixes
* refactoring
* refactoring
* nullopt fix
* alpine segfault fix
* refactoring
* Merge remote-tracking branch 'main/master' into data_def_length

+++ language_server/src/main.cpp
@@ -58,7 +58,7 @@ int main(int argc, char ** argv) {
 			}
 
 			asio::io_service io_service_;
-			asio::ip::tcp::acceptor acceptor_(io_service_, asio::ip::tcp::endpoint(asio::ip::tcp::v4(), port));
+			asio::ip::tcp::acceptor acceptor_(io_service_, asio::ip::tcp::endpoint(asio::ip::tcp::v4(), (uint16_t) port));
 			asio::ip::tcp::socket socket_(io_service_);
 			asio::ip::tcp::iostream stream;
 			acceptor_.accept(stream.socket());
